Fix parent dir create
Previously, if the parent dir path starts with a "/", it may be ignored.
As a result, a useless copy of the dir hierarchy will be created based
on the current working directory of the process.

The fix checks first whether there is a "/" prefix, and if there is, the
result path will be prefixed with a "/" to preserve the original root of
the file path.